requirements.txt
streamlit
langchain
langchain-community
langchain-openai
faiss-cpu
sentence-transformers
duckduckgo-search
python-dotenv
openai
groq
google-generativeai